get video time in application closed

+++ app/src/main/java/com/mkandeel/youtubeplayer/VideoStateChanged.kt
@@ -1,5 +1,5 @@
 package com.mkandeel.youtubeplayer
 
 interface VideoStateChanged {
-    fun onVideoStateChanged(isPaused: Boolean, currentTime: Float)
+    fun onVideoStateChanged(isPaused: Boolean, currentTime: Int)
 }
